[♻️ Refactor] 차트 아이템 구조 변경 (#17)
