+++ components/board/ChartBoard/BoardItem/CompareChart.tsx
