+++ components/board/ChartBoard/BoardItem/YearlyChart.tsx
