+++ components/board/ChartBoard/ChartIndexBoard.tsx
@@ -1,10 +1,10 @@
 import ChartTapButton from "@/components/button/ChartTapButton";
 import React, { useState } from "react";
 import { ScrollView, View } from "react-native";
-import CompareChart from "./CompareChart";
-import MonthlyChart from "./MonthlyChart";
-import WeeklyChart from "./WeeklyChart";
-import YearlyChart from "./YearlyChart";
+import CompareChart from "./BoardItem/CompareChart";
+import MonthlyChart from "./BoardItem/MonthlyChart";
+import WeeklyChart from "./BoardItem/WeeklyChart";
+import YearlyChart from "./BoardItem/YearlyChart";
 
 const ChartIndexBoard = () => {
   const [selectedChart, setSelectedChart] = useState("weekly");
@@ -27,7 +27,7 @@ const ChartIndexBoard = () => {
 
   return (
     <ScrollView className="mb-4 flex-1 bg-gray-100">
-      <View className=" m-4 min-h-[400px] justify-around rounded-lg bg-white p-4">
+      <View className=" m-4 min-h-[390px] justify-around rounded-lg bg-white p-4">
         <View className="flex-row justify-around">
           <ChartTapButton
             handleSelectChart={handleSelectChart}
